refactor(app): move base type definition inside types folder and move helper function to new file
