+++ packages/app/src/components/dialog/settings/DebugOptionsPage.tsx
@@ -5,7 +5,7 @@ import { toast } from 'sonner'
 import { useShallow } from 'zustand/shallow'
 
 import { useEditorStore } from '@/stores/editorStore'
-import type { LogLevel } from '@/types'
+import type { LogLevel } from '@/types/base'
 
 const DebugOptionsPage: FC = () => {
   const { t } = useTranslation()
